Move tile layers to editor state
