+++ src/hooks/useEditorState.tsx
@@ -6,11 +6,16 @@ export interface EditorState {
     foregroundVisible: boolean;
     leftClickTileId: number;
     rightClickTileId: number;
+    foregroundTiles: Array<number>;
+    backgroundTiles: Array<number>;
     setLeftClickPressed(pressed: boolean): void;
     setRightClickPressed(pressed: boolean): void;
     toggleForegroundVisible(): void;
     setLeftClickTileId(tileId: number): void;
     setRightClickTileId(tileId: number): void;
+    placeForegroundTileAtIndex(tileId: number, index: number): void,
+    placeBackgroundTileAtIndex(tileId: number, index: number): void,
+    pickTile(tileId: number): void,
 }
 
 function getEmptyEditorState(): EditorState {
@@ -20,12 +25,21 @@ function getEmptyEditorState(): EditorState {
         foregroundVisible: true,
         leftClickTileId: 0,
         rightClickTileId: 0,
+        foregroundTiles: getBlankTileGrid(),
+        backgroundTiles: getBlankTileGrid(),
         setLeftClickPressed: (b: boolean) => { throw new Error("Editor State Context not found") },
         setRightClickPressed: (b: boolean) => { throw new Error("Editor State Context not found") },
         toggleForegroundVisible: () => { throw new Error("Editor State Context not found") },
         setLeftClickTileId: (tileId: number) => { throw new Error("Editor State Context not found") },
-        setRightClickTileId: (tileId: number) => { throw new Error("Editor State Context not found") }
+        setRightClickTileId: (tileId: number) => { throw new Error("Editor State Context not found") },
+        placeForegroundTileAtIndex: (tileId: number, index: number) => { throw new Error("Editor State Context not found") },
+        placeBackgroundTileAtIndex: (tileId: number, index: number) => { throw new Error("Editor State Context not found") },
+        pickTile: (tileId: number) => { throw new Error("Editor State Context not found") },
     }
+}
+
+function getBlankTileGrid(): Array<number> {
+    return Array(20 * 15).fill(0);
 }
 
 export const EditorStateContext = createContext(getEmptyEditorState());
@@ -40,6 +54,26 @@ export function EditorStateProvider(props: {children: any}) {
     const [foregroundVisible, setForegroundVisible] = useState(true);
     const [leftClickTileId, setLeftClickTileId] = useState(0);
     const [rightClickTileId, setRightClickTileId] = useState(0);
+    const [backgroundTiles, setBackgroundTiles] = useState(getBlankTileGrid());
+    const [foregroundTiles, setForegroundTiles] = useState(getBlankTileGrid());
+
+    function placeBackgroundTileAtIndex(index: number, button: number) {
+        const tileToAdd = button === 0 ? leftClickTileId : rightClickTileId;
+        setBackgroundTiles((bt) => {
+            return [...bt.slice(0, index), tileToAdd, ...bt.slice(index + 1)];
+        });
+    }
+
+    function placeForegroundTileAtIndex(index: number, button: number) {
+        const tileToAdd = button === 0 ? leftClickTileId : rightClickTileId;
+        setForegroundTiles((ft) => {
+            return [...ft.slice(0, index), tileToAdd, ...ft.slice(index + 1)];
+        });
+    }
+
+    function pickTile(tileId: number) {
+        setLeftClickTileId(tileId);
+    }
 
     const value = useMemo(() => ({
         leftClickPressed,
@@ -47,12 +81,17 @@ export function EditorStateProvider(props: {children: any}) {
         foregroundVisible,
         leftClickTileId,
         rightClickTileId,
+        foregroundTiles,
+        backgroundTiles,
         setLeftClickPressed,
         setRightClickPressed,
         toggleForegroundVisible: () => setForegroundVisible(fv => !fv),
         setLeftClickTileId: (id: number) => setLeftClickTileId(id),
-        setRightClickTileId: (id: number) => setRightClickTileId(id)
-    }), [leftClickPressed, rightClickPressed, foregroundVisible, leftClickTileId, rightClickTileId]);
+        setRightClickTileId: (id: number) => setRightClickTileId(id),
+        placeForegroundTileAtIndex,
+        placeBackgroundTileAtIndex,
+        pickTile,
+    }), [leftClickPressed, rightClickPressed, foregroundVisible, leftClickTileId, rightClickTileId, foregroundTiles, backgroundTiles]);
 
     return (
         <EditorStateContext.Provider value={value}>
